chore: refactor data imports and update TypeScript configuration
- Updated import paths to reference the new `data.ts` file instead of the previous `data.js`, ensuring consistency across the codebase.
- Enhanced the TypeScript configuration by extending the base configuration and adding specific compiler options, improving build output and type checking.
- Introduced a new `data.ts` file that includes utility functions and constants, streamlining data handling and improving code organization.

These changes contribute to a cleaner and more maintainable codebase, enhancing the overall development experience.
