+++ docs/src/components/SocialIcons.astro
@@ -1,7 +1,7 @@
 ---
 import SocialIcons_ from '@astrojs/starlight/components/SocialIcons.astro'
 import { liveStoreVersion } from '@livestore/common'
-import { IS_MAIN_BRANCH } from '../../data.js'
+import { IS_MAIN_BRANCH } from '../data/data.ts'
 
 const devUrl = 'https://dev.docs.livestore.dev'
 const latestUrl = 'https://docs.livestore.dev'
