+++ docs/src/content/docs/examples/cloudflare-adapter.mdx
@@ -6,7 +6,7 @@ description: LiveStore examples built with the Cloudflare adapter for edge compu
 import { Card, CardGrid } from '@astrojs/starlight/components';
 import { Image } from 'astro:assets';
 import { cloudflareExamples, getExampleDemoLinks } from '../../../data/examples.ts';
-import { getBranchName } from '../../../../data.js';
+import { getBranchName } from '../../../data/data.ts';
 
 # Cloudflare Durable Objects Examples
