+++ docs/src/content/docs/examples/expo-adapter.mdx
@@ -6,7 +6,7 @@ description: LiveStore examples built with the Expo adapter for React Native mob
 import { Card, CardGrid } from '@astrojs/starlight/components';
 import { Image } from 'astro:assets';
 import { expoExamples, getExampleDemoLinks } from '../../../data/examples.ts';
-import { getBranchName } from '../../../../data.js';
+import { getBranchName } from '../../../data/data.ts';
 
 # Expo Adapter Examples
