+++ docs/src/content/docs/examples/index.mdx
@@ -5,7 +5,7 @@ description: Explore LiveStore example applications across different platforms a
 
 import { Card, CardGrid, LinkCard } from '@astrojs/starlight/components';
 import { webExamples, nodeExamples, expoExamples, cloudflareExamples } from '../../../data/examples.ts';
-import { getBranchName } from '../../../../data.js';
+import { getBranchName } from '../../../data/data.ts';
 
 # Example Applications
