+++ docs/src/content/docs/examples/node-adapter.mdx
@@ -6,7 +6,7 @@ description: LiveStore examples built with the Node adapter for server-side envi
 import { Card, CardGrid } from '@astrojs/starlight/components';
 import { Image } from 'astro:assets';
 import { nodeExamples, getExampleDemoLinks } from '../../../data/examples.ts';
-import { getBranchName } from '../../../../data.js';
+import { getBranchName } from '../../../data/data.ts';
 
 # Node Adapter Examples
