+++ docs/src/content/docs/examples/web-adapter.mdx
@@ -6,7 +6,7 @@ description: LiveStore examples built with the Web adapter for browser environme
 import { Card, CardGrid } from '@astrojs/starlight/components';
 import { Image } from 'astro:assets';
 import { webExamples, getExampleDemoLinks } from '../../../data/examples.ts';
-import { getBranchName } from '../../../../data.js';
+import { getBranchName } from '../../../data/data.ts';
 
 # Web Adapter Examples
