+++ docs/src/content/docs/getting-started/expo.mdx
@@ -6,7 +6,7 @@ sidebar:
 ---
 
 import { Code, Steps, Tabs, TabItem } from '@astrojs/starlight/components'
-import { makeTiged, versionNpmSuffix } from '../../../../data.js'
+import { makeTiged, versionNpmSuffix } from '../../../data/data.ts'
 import { MIN_NODE_VERSION } from '@local/shared'
 import babelConfigCode from '../../../../../examples/expo-todomvc-sync-cf/babel.config.js?raw'
 import metroConfigCode from '../../../../../examples/expo-todomvc-sync-cf/metro.config.js?raw'
