+++ docs/src/content/docs/getting-started/node.mdx
@@ -6,7 +6,7 @@ sidebar:
 ---
 
 import { Steps, Tabs, TabItem, Code } from '@astrojs/starlight/components';
-import { makeTiged } from '../../../../data.js'
+import { makeTiged } from '../../../data/data.ts'
 
 ## Minimal example
