+++ docs/src/content/docs/getting-started/react-web.mdx
@@ -7,7 +7,7 @@ sidebar:
 ---
 
 import { Steps, Tabs, TabItem, Code } from '@astrojs/starlight/components';
-import { makeTiged, versionNpmSuffix } from '../../../../data.js'
+import { makeTiged, versionNpmSuffix } from '../../../data/data.ts'
 import { MIN_NODE_VERSION } from '@local/shared'
 import viteConfigCode from '../../../../../examples/web-todomvc-sync-cf/vite.config.js?raw'
 import schemaCode from '../../../../../examples/web-todomvc-sync-cf/src/livestore/schema.ts?snippet'
