+++ docs/src/content/docs/getting-started/vue.mdx
@@ -7,7 +7,7 @@ sidebar:
 ---
 
 import { Steps, Tabs, TabItem, Code } from '@astrojs/starlight/components';
-import { makeTiged, versionNpmSuffix } from '../../../../data.js'
+import { makeTiged, versionNpmSuffix } from '../../../data/data.ts'
 import { MIN_NODE_VERSION } from '@local/shared'
 import schemaCode from '../../_assets/code/getting-started/vue/livestore/schema.ts?snippet'
 import workerCode from '../../_assets/code/getting-started/vue/livestore/livestore.worker.ts?snippet'
