+++ docs/src/content/docs/index.mdx
@@ -3,7 +3,7 @@ title: Docs
 ---
 
 import { CardGrid, LinkCard } from '@astrojs/starlight/components';
-import { officeHours } from '../../../data.js'
+import { officeHours } from '../../data/data.ts'
 import { liveStoreVersion } from '@livestore/common'
 import NpmLink from '../../components/NpmLink.astro'
