+++ docs/src/content/docs/misc/community.mdx
@@ -5,7 +5,7 @@ sidebar:
 ---
 
 import { DISCORD_INVITE_URL } from '@local/shared'
-import { officeHours } from '../../../../data.js'
+import { officeHours } from '../../../data/data.ts'
 
 ## Discord
