+++ docs/src/pages/og/[...slug].ts
@@ -22,16 +22,14 @@ export const { getStaticPaths, GET } = OGImageRoute({
   // as the file is named `[...slug].ts`.
   param: 'slug',
   // Define a function called for each page to customize the generated image.
-  getImageOptions: (_id: string, page: (typeof pages)[string]) => {
-    return {
-      // Use the page title and description as the image title and description.
-      title: page.data.title,
-      description: page.data.description,
-      // Customize various colors and add a border.
-      bgGradient: [[24, 24, 27]],
-      border: { color: [63, 63, 70], width: 20 },
-      padding: 60,
-      logo: { path: './src/logo.png', size: [180] },
-    }
-  },
+  getImageOptions: (_id: string, page: (typeof pages)[string]) => ({
+    // Use the page title and description as the image title and description.
+    title: page.data.title,
+    description: page.data.description ?? '',
+    // Customize various colors and add a border.
+    bgGradient: [[24, 24, 27]],
+    border: { color: [63, 63, 70], width: 20 },
+    padding: 60,
+    logo: { path: './src/logo.png', size: [180] },
+  }),
 })
